refactor: unify navigation system across all UI components
  - Create core/navigation.py with consolidated navigation utilities
  - Remove duplicate questionary styling from ui.py and BaseUIComponent
  - Update all UI classes (ClusterUI, ServiceUI, TaskUI) to use shared functions
  - Fix imports to use navigation utilities from correct location
  - Eliminate ~50 lines of duplicate navigation code
  - Standardize navigation patterns and error handling across components
